fix: Update helper functions
- extractDataFromBody
The request body passed to the endpoint is a formdata and certain fields
should not be passed to the database. This function handles that.

- canSaveToDB
The condition for submitting the application is different as it has to
be confirmed that all other stages has been completed.

- initialDriveFile
This was created to cater to the schema for resume and coverLetter. In
the absence of the optional coverLetter, a value that aligns with schema
is passed.

- uploadFile
This is the function that uploads the file to google drive.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>